reformatted some components to use Es6 classes. Fixed casing on some files. Moved some components into folders
