+++ app/components/ProfileChart/ProfileChart.js
