--- app/components/event.scss
@@ -1,8 +1,0 @@
-.box {
-  border: 10px solid #ccc;
-  padding-top: 10px;
-  padding: 15px;
-  margin: 0; 
-}
-
-
