+++ app/router.js
@@ -5,7 +5,7 @@ import App from 'views/App'
 import ProfileView from 'views/profile/ProfileView'
 import HomeView from 'views/home/HomeView'
 import SearchView from 'views/search/SearchView'
-import LoginView from 'views/login/loginView'
+import LoginView from 'views/login/LoginView'
 import RegisterView from 'views/register/RegisterView'
 import ApprovalView from 'views/approval/ApprovalView'
